Password validation & database connection mechs
This commit changes a few things, including how the database is
connected to.

The database connector is now placed in development_db_password.inc.php,
and not in the files themselves.

This change also changes how password validation is done, and it should,
theoretically, work.

Also, removed $this->xxx unneeded variables and swapped them with $xxx
ones.